+++ src/main/java/net/sf/jabref/logic/integrity/IntegrityCheck.java
@@ -51,6 +51,8 @@ public class IntegrityCheck {
         result.addAll(new BracketChecker(FieldName.TITLE).check(entry));
         result.addAll(new YearChecker().check(entry));
         result.addAll(new EditionChecker(bibDatabaseContext).check(entry));
+        result.addAll(new NoteChecker(bibDatabaseContext).check(entry));
+        result.addAll(new HowpublishedChecker(bibDatabaseContext).check(entry));
         result.addAll(new UrlChecker().check(entry));
         result.addAll(new FileChecker(bibDatabaseContext, fileDirectoryPreferences).check(entry));
         result.addAll(new TypeChecker().check(entry));